Remove multipass
