+++ examples/glfx/index.html
@@ -11,6 +11,7 @@
       crossorigin="anonymous"></script>
 
     <script type="text/javascript" src="./dist/lumagl.js"></script>
+    <script type="text/javascript" src="multipass.js"></script>
     <script type="text/javascript" src="canvas.js"></script>
     <script type="text/javascript" src="app.js"></script>
