--- modules/core/test/multipass/index.js
@@ -1,11 +1,0 @@
-import './pass.spec';
-import './multi-pass-render.spec';
-import './clear-mask-pass.spec';
-import './composite-pass.spec';
-import './copy-pass.spec';
-import './mask-pass.spec';
-import './render-pass.spec';
-import './stencil-pass.spec';
-import './texture-pass.spec';
-import './clear-pass.spec';
-import './shader-module-pass.spec';
