+++ modules/core/test/webgl-dependent-tests.js
@@ -1,2 +1,1 @@
 import './lib';
-import './multipass';
