--- modules/effects/src/experimental/passes/ssao-pass.md
@@ -1,15 +1,0 @@
-# SSAOPass (Experimental)
-
-A Screen Space Ambient Occlusion Render Pass
-
-References:
-
-* Based on SSAO shader module
-
-
-## Properties
-
-* `radius` - Ambient occlusion shadow radius (numeric value).
-* `onlyAO` - Display only ambient occlusion result (boolean value).
-* `aoClamp` - Ambient occlusion clamp (numeric value).
-* `lumInfluence` - Pixel luminosity influence in AO calculation (numeric value).
