+++ modules/effects/src/index.js
@@ -31,5 +31,3 @@ export {default as fxaa} from './shader-modules/fxaa/fxaa';
 // experimental shader modules and passes
 export {default as _depth} from './experimental/shader-modules/depth';
 export {default as _convolution} from './experimental/shader-modules/convolution';
-export {default as _OutlinePass} from './experimental/passes/outline-pass';
-export {default as _SSAOPass} from './experimental/passes/ssao-pass';
